🚚 refactor: Update details form components folder structure
