+++ src/views/DetailsForm/components/DestinationInput/DestinationInput.tsx
