+++ src/views/DetailsForm/components/IndicationsInput/IndicationsInput.tsx
